chore: Move main.go to root
