+++ .goreleaser.yml
@@ -5,7 +5,6 @@ before:
 builds:
 - env:
   - CGO_ENABLED=0
-  main: ./cmd/lefthook/
   goos:
     - linux
     - darwin
