+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/evilmartians/lefthook/cmd"
+
+func main() {
+	cmd.Lefthook()
+}
